@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: ./api.go
+//
+// Generated by this command:
+//
+//	mockgen -destination=./mock/api.go -source=./api.go Provider
+//
 
 // Package mock_api is a generated GoMock package.
 package mock_api
@@ -18,6 +23,7 @@ import (
 type MockApi struct {
 	ctrl     *gomock.Controller
 	recorder *MockApiMockRecorder
+	isgomock struct{}
 }
 
 // MockApiMockRecorder is the mock recorder for MockApi.
@@ -46,7 +52,7 @@ func (m *MockApi) Listen(port string) error {
 }
 
 // Listen indicates an expected call of Listen.
-func (mr *MockApiMockRecorder) Listen(port interface{}) *gomock.Call {
+func (mr *MockApiMockRecorder) Listen(port any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Listen", reflect.TypeOf((*MockApi)(nil).Listen), port)
 }
@@ -54,7 +60,7 @@ func (mr *MockApiMockRecorder) Listen(port interface{}) *gomock.Call {
 // Test mocks base method.
 func (m *MockApi) Test(req *http.Request, msTimeout ...int) (*http.Response, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{req}
+	varargs := []any{req}
 	for _, a := range msTimeout {
 		varargs = append(varargs, a)
 	}
@@ -65,9 +71,9 @@ func (m *MockApi) Test(req *http.Request, msTimeout ...int) (*http.Response, err
 }
 
 // Test indicates an expected call of Test.
-func (mr *MockApiMockRecorder) Test(req interface{}, msTimeout ...interface{}) *gomock.Call {
+func (mr *MockApiMockRecorder) Test(req any, msTimeout ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{req}, msTimeout...)
+	varargs := append([]any{req}, msTimeout...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Test", reflect.TypeOf((*MockApi)(nil).Test), varargs...)
 }
 
@@ -75,6 +81,7 @@ func (mr *MockApiMockRecorder) Test(req interface{}, msTimeout ...interface{}) *
 type MockProvider struct {
 	ctrl     *gomock.Controller
 	recorder *MockProviderMockRecorder
+	isgomock struct{}
 }
 
 // MockProviderMockRecorder is the mock recorder for MockProvider.
@@ -95,15 +102,16 @@ func (m *MockProvider) EXPECT() *MockProviderMockRecorder {
 }
 
 // AdjustEndpoints mocks base method.
-func (m *MockProvider) AdjustEndpoints(endpoints []*endpoint.Endpoint) []*endpoint.Endpoint {
+func (m *MockProvider) AdjustEndpoints(endpoints []*endpoint.Endpoint) ([]*endpoint.Endpoint, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "AdjustEndpoints", endpoints)
 	ret0, _ := ret[0].([]*endpoint.Endpoint)
-	return ret0
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // AdjustEndpoints indicates an expected call of AdjustEndpoints.
-func (mr *MockProviderMockRecorder) AdjustEndpoints(endpoints interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) AdjustEndpoints(endpoints any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AdjustEndpoints", reflect.TypeOf((*MockProvider)(nil).AdjustEndpoints), endpoints)
 }
@@ -117,16 +125,16 @@ func (m *MockProvider) ApplyChanges(ctx context.Context, changes *plan.Changes) 
 }
 
 // ApplyChanges indicates an expected call of ApplyChanges.
-func (mr *MockProviderMockRecorder) ApplyChanges(ctx, changes interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) ApplyChanges(ctx, changes any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ApplyChanges", reflect.TypeOf((*MockProvider)(nil).ApplyChanges), ctx, changes)
 }
 
 // GetDomainFilter mocks base method.
-func (m *MockProvider) GetDomainFilter() endpoint.DomainFilter {
+func (m *MockProvider) GetDomainFilter() endpoint.DomainFilterInterface {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetDomainFilter")
-	ret0, _ := ret[0].(endpoint.DomainFilter)
+	ret0, _ := ret[0].(endpoint.DomainFilterInterface)
 	return ret0
 }
 
@@ -146,7 +154,7 @@ func (m *MockProvider) Records(ctx context.Context) ([]*endpoint.Endpoint, error
 }
 
 // Records indicates an expected call of Records.
-func (mr *MockProviderMockRecorder) Records(ctx interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) Records(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Records", reflect.TypeOf((*MockProvider)(nil).Records), ctx)
 }
