@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: ./http_middleware.go
+//
+// Generated by this command:
+//
+//	mockgen -destination=./mock/http_middleware.go -source=./http_middleware.go HttpApiMetrics
+//
 
 // Package mock_metrics is a generated GoMock package.
 package mock_metrics
@@ -14,6 +19,7 @@ import (
 type MockHttpApiMetrics struct {
 	ctrl     *gomock.Controller
 	recorder *MockHttpApiMetricsMockRecorder
+	isgomock struct{}
 }
 
 // MockHttpApiMetricsMockRecorder is the mock recorder for MockHttpApiMetrics.
@@ -64,7 +70,7 @@ func (m *MockHttpApiMetrics) CollectRequest(method, path string, statusCode int)
 }
 
 // CollectRequest indicates an expected call of CollectRequest.
-func (mr *MockHttpApiMetricsMockRecorder) CollectRequest(method, path, statusCode interface{}) *gomock.Call {
+func (mr *MockHttpApiMetricsMockRecorder) CollectRequest(method, path, statusCode any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CollectRequest", reflect.TypeOf((*MockHttpApiMetrics)(nil).CollectRequest), method, path, statusCode)
 }
@@ -76,7 +82,7 @@ func (m *MockHttpApiMetrics) CollectRequestContentLength(method, path string, co
 }
 
 // CollectRequestContentLength indicates an expected call of CollectRequestContentLength.
-func (mr *MockHttpApiMetricsMockRecorder) CollectRequestContentLength(method, path, contentLength interface{}) *gomock.Call {
+func (mr *MockHttpApiMetricsMockRecorder) CollectRequestContentLength(method, path, contentLength any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CollectRequestContentLength", reflect.TypeOf((*MockHttpApiMetrics)(nil).CollectRequestContentLength), method, path, contentLength)
 }
@@ -88,7 +94,7 @@ func (m *MockHttpApiMetrics) CollectRequestDuration(method, path string, duratio
 }
 
 // CollectRequestDuration indicates an expected call of CollectRequestDuration.
-func (mr *MockHttpApiMetricsMockRecorder) CollectRequestDuration(method, path, duration interface{}) *gomock.Call {
+func (mr *MockHttpApiMetricsMockRecorder) CollectRequestDuration(method, path, duration any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CollectRequestDuration", reflect.TypeOf((*MockHttpApiMetrics)(nil).CollectRequestDuration), method, path, duration)
 }
@@ -100,7 +106,7 @@ func (m *MockHttpApiMetrics) CollectRequestResponseSize(method, path string, con
 }
 
 // CollectRequestResponseSize indicates an expected call of CollectRequestResponseSize.
-func (mr *MockHttpApiMetricsMockRecorder) CollectRequestResponseSize(method, path, contentLength interface{}) *gomock.Call {
+func (mr *MockHttpApiMetricsMockRecorder) CollectRequestResponseSize(method, path, contentLength any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CollectRequestResponseSize", reflect.TypeOf((*MockHttpApiMetrics)(nil).CollectRequestResponseSize), method, path, contentLength)
 }
