@@ -2,6 +2,6 @@ package stackitprovider
 
 import "sigs.k8s.io/external-dns/endpoint"
 
-func (d *StackitDNSProvider) GetDomainFilter() endpoint.DomainFilter {
-	return d.domainFilter
+func (d *StackitDNSProvider) GetDomainFilter() endpoint.DomainFilterInterface {
+	return &d.domainFilter
 }
