@@ -18,7 +18,13 @@ func (w webhook) AdjustEndpoints(ctx *fiber.Ctx) error {
 		return ctx.Status(fiber.StatusBadRequest).SendString(err.Error())
 	}
 
-	pve = w.provider.AdjustEndpoints(pve)
+	pve, err = w.provider.AdjustEndpoints(pve)
+	if err != nil {
+		w.logger.Error("Error adjusting endpoints", zap.String(logFieldError, err.Error()))
+		ctx.Response().Header.Set(contentTypeHeader, contentTypePlaintext)
+
+		return ctx.Status(fiber.StatusInternalServerError).SendString(err.Error())
+	}
 
 	w.logger.Debug("adjusted endpoints", zap.String("endpoints", fmt.Sprintf("%v", pve)))
 
